Move TabloidRedis class to RDB test module
Now that the memory backend can also produce dumps (#77), there is very
limited use for TabloidRedis. It is only needed when testing the
saving and loading of RDB files via the Redis backend interface.

This commit is a straight copy 'n' paste of the class to make it easy
to track the cross-file changes via git. The next commit will patch
things up a bit.
